ai_model/requirements.txt
flask==2.0.1
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.2.0+cu121
torchvision==0.17.0+cu121
diffusers==0.25.0
transformers==4.37.2
accelerate==0.27.2
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
werkzeug==2.0.3
huggingface-hub==0.20.3
numpy==1.24.3
flask-cors==4.0.0
watchdog==3.0.0
